Rename app to org.mozilla.webmaker
