+++ app/build.gradle
@@ -5,7 +5,7 @@ android {
     buildToolsVersion "21.1.2"
 
     defaultConfig {
-        applicationId "mozilla.org.webmaker"
+        applicationId "org.mozilla.webmaker"
         minSdkVersion 14
         targetSdkVersion 22
         versionCode 1
