+++ app/src/androidTest/java/org/mozilla/webmaker/ApplicationTest.java
@@ -1,0 +1,13 @@
+package org.mozilla.webmaker;
+
+import android.app.Application;
+import android.test.ApplicationTestCase;
+
+/**
+ * <a href="http://d.android.com/tools/testing/testing_android.html">Testing Fundamentals</a>
+ */
+public class ApplicationTest extends ApplicationTestCase<Application> {
+    public ApplicationTest() {
+        super(Application.class);
+    }
+}
