--- app/src/main/java/mozilla/org/webmaker/BaseActivity.java
@@ -1,9 +1,0 @@
-package mozilla.org.webmaker;
-
-import android.app.Activity;
-
-public class BaseActivity extends Activity {
-    public void goBack() {
-        super.onBackPressed();
-    }
-}
