--- app/src/main/java/mozilla/org/webmaker/activity/Page.java
@@ -1,10 +1,0 @@
-package mozilla.org.webmaker.activity;
-
-import mozilla.org.webmaker.R;
-import mozilla.org.webmaker.WebmakerActivity;
-
-public class Page extends WebmakerActivity {
-    public Page() {
-        super("page", R.id.page_layout, R.layout.page_layout, R.menu.menu_page);
-    }
-}
