--- app/src/main/java/mozilla/org/webmaker/activity/Play.java
@@ -1,10 +1,0 @@
-package mozilla.org.webmaker.activity;
-
-import mozilla.org.webmaker.R;
-import mozilla.org.webmaker.WebmakerActivity;
-
-public class Play extends WebmakerActivity {
-    public Play() {
-        super("project", R.id.page_layout, R.layout.page_layout, R.menu.menu_page);
-    }
-}
