--- app/src/main/java/mozilla/org/webmaker/activity/ProjectSettings.java
@@ -1,10 +1,0 @@
-package mozilla.org.webmaker.activity;
-
-import mozilla.org.webmaker.R;
-import mozilla.org.webmaker.WebmakerActivity;
-
-public class ProjectSettings extends WebmakerActivity {
-    public ProjectSettings() {
-        super("project-settings", R.id.page_layout, R.layout.page_layout, R.menu.menu_page);
-    }
-}
