--- app/src/main/java/mozilla/org/webmaker/router/exception/NoContextProvidedException.java
@@ -1,10 +1,0 @@
-package mozilla.org.webmaker.router.exception;
-
-public class NoContextProvidedException extends RuntimeException {
-
-    private static final long serialVersionUID = -1381427067387547157L;
-
-    public NoContextProvidedException(String message) {
-        super(message);
-    }
-}
