--- app/src/main/java/mozilla/org/webmaker/router/helper/RouterParams.java
@@ -1,8 +1,0 @@
-package mozilla.org.webmaker.router.helper;
-
-import java.util.Map;
-
-public class RouterParams {
-    public RouterOptions routerOptions;
-    public Map<String, String> openParams;
-}
