+++ app/src/main/java/org/mozilla/webmaker/BaseActivity.java
@@ -1,0 +1,9 @@
+package org.mozilla.webmaker;
+
+import android.app.Activity;
+
+public class BaseActivity extends Activity {
+    public void goBack() {
+        super.onBackPressed();
+    }
+}
