+++ app/src/main/java/org/mozilla/webmaker/activity/ProjectSettings.java
@@ -1,0 +1,10 @@
+package org.mozilla.webmaker.activity;
+
+import org.mozilla.webmaker.R;
+import org.mozilla.webmaker.WebmakerActivity;
+
+public class ProjectSettings extends WebmakerActivity {
+    public ProjectSettings() {
+        super("project-settings", R.id.page_layout, R.layout.page_layout, R.menu.menu_page);
+    }
+}
