+++ app/src/main/java/org/mozilla/webmaker/router/helper/RouterParams.java
@@ -1,0 +1,8 @@
+package org.mozilla.webmaker.router.helper;
+
+import java.util.Map;
+
+public class RouterParams {
+    public RouterOptions routerOptions;
+    public Map<String, String> openParams;
+}
