+++ app/src/main/res/layout/page_layout.xml
@@ -1,7 +1,7 @@
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="mozilla.org.webmaker.activity.Page"
+    tools:context="org.mozilla.webmaker.activity.Page"
     android:id="@+id/page_layout">
 
 </RelativeLayout>
