+++ app/src/main/res/menu/menu_project.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
       xmlns:tools="http://schemas.android.com/tools" xmlns:app="http://schemas.android.com/apk/res-auto"
-      tools:context="mozilla.org.webmaker.activity.Project">
+      tools:context="org.mozilla.webmaker.activity.Project">
 
     <item android:id="@+id/action_play"
           android:title="@string/action_play"
